Release 1.0.9 improvements OverAgressive confirmation request.

+++ src/syncModule.ts
@@ -861,15 +861,17 @@ export class SyncMan {
 			const reallyDeletedTickTickTasks = deletedTickTickTasks.filter(task => deletedTasks.some(t => t.taskId === task.id));
 			// this.dumpArray('== reallyDeletedTickTickTasks deleted from TickTick:', reallyDeletedTickTickTasks);
 
-			const taskTitlesForConfirmation = reallyDeletedTickTickTasks.map((task: ITask) => task.id)
+			if (reallyDeletedTickTickTasks.length > 0) {
+				const taskTitlesForConfirmation = reallyDeletedTickTickTasks.map((task: ITask) => task.id)
 
-			const bConfirm = await  this.confirmDeletion(taskTitlesForConfirmation, "tasks deleted from TickTick");
+				const bConfirm = await this.confirmDeletion(taskTitlesForConfirmation, "tasks deleted from TickTick");
 
-			if (bConfirm) {
-				for (const task of reallyDeletedTickTickTasks) {
-					await this.plugin.fileOperation?.deleteTaskFromFile(task);
-					await this.plugin.cacheOperation?.deleteTaskFromCache(task.id)
-					bModifiedFileSystem = true;
+				if (bConfirm) {
+					for (const task of reallyDeletedTickTickTasks) {
+						await this.plugin.fileOperation?.deleteTaskFromFile(task);
+						await this.plugin.cacheOperation?.deleteTaskFromCache(task.id)
+						bModifiedFileSystem = true;
+					}
 				}
 			}
